Check delete request id matches blog id

+++ part4/bloglist/controller/blogs.js
@@ -57,8 +57,22 @@ blogsRouter.delete('/:id', async (request, response) => {
     })
   }
 
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).json({
+      error: 'blog not found'
+    })
+  }
+
+  if (user._id.toString() === blog.user.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(400).json({
+      error: 'the blog could not be deleted'
+    })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
